+++ src/components/Slices/Content.js
@@ -1,7 +1,7 @@
 import React from "react"
 
-const Content = ({ html }) => (
-  <div className="text-4xl" dangerouslySetInnerHTML={{ __html: html }} />
+const Content = ({ className, html }) => (
+  <div className={className} dangerouslySetInnerHTML={{ __html: html }} />
 )
 
 export default Content
